🚚 Moved and renamed `IntegrationsTagList` component
